FileStatusクラスの名称変更に合わせてテスト名称も変更
